pyproject.toml
[tool.poetry]
name = "gitgus"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
gitpython = "^3.1.31"
pygithub = "^1.58.0"
toml = "^0.10.2"
slack-sdk = "^3.20.1"
rich = "^13.3.2"
dateparser = "^1.1.8"
markdown = "^3.4.3"
python-jenkins = "^1.8.0"
pydantic = "^2.2.1"
typer = "^0.9.0"
simple-salesforce = "^1.12.4"
stringcase = "^1.2.0"
sh = "^2.0.6"
jinja2 = "^3.1.2"


[tool.poetry.scripts]
gitgus = 'gitgus.main:app'


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.pytest.ini_options]
log_cli = true
timeout = 6

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
flake8-pytest-style = "^1.7.2"
pytest-cov = "^4.0.0"
black = "^23.1.0"
flakeheaven = "^3.2.1"
pyfakefs = "^5.2.2"
pytest-timeout = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flakeheaven]
format = "grouped"
max_line_length = 120
show_source = true
exclude = ["venv/*"]

[tool.flakeheaven.plugins]
pyflakes = ["+*", "-F401", "-F821"] # F401: module imported but unused, annoying with __init__.py, F821: undefined name (false positive)
flake8-pytest-style = ["+*"]
mccabe = ["+*"]
pycodestyle = ["+*", "-E501", "-W503", "-E203"] # E501: line too long, W503: line break before binary operator,
pylint = ["+*"]
